# Optimise NEON assembly for get_sad for width 16