Reduces execution time for this function by between 18% and 40%.
The most commonly used block size, 16x16, is improved by 29%.

<details>

```
get_sad/(BLOCK_16X8, 8) time:   [10.095 ns 10.104 ns 10.111 ns]                                     
                        change: [-25.513% -25.484% -25.457%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 25 outliers among 100 measurements (25.00%)
  7 (7.00%) low severe
  1 (1.00%) low mild
  5 (5.00%) high mild
  12 (12.00%) high severe
get_sad/(BLOCK_16X16, 8)                                                                             
                        time:   [15.758 ns 15.769 ns 15.780 ns]
                        change: [-30.763% -30.703% -30.644%] (p = 0.00 < 0.05)
                        Performance has improved.
get_sad/(BLOCK_16X32, 8)                                                                             
                        time:   [26.581 ns 26.586 ns 26.592 ns]
                        change: [-35.741% -35.705% -35.668%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  12 (12.00%) high severe
get_sad/(BLOCK_16X4, 8) time:   [6.9992 ns 7.0004 ns 7.0019 ns]                                     
                        change: [-19.098% -18.976% -18.842%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low mild
  5 (5.00%) high mild
  3 (3.00%) high severe
get_sad/(BLOCK_16X64, 8)                                                                             
                        time:   [47.702 ns 47.710 ns 47.720 ns]
                        change: [-38.632% -38.550% -38.465%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 17 outliers among 100 measurements (17.00%)
  5 (5.00%) high mild
  12 (12.00%) high severe
```
</details>